Implement movement and restarting. Fixes #6
Ball now moves in random direction at a random speed on game start. Pressing 'R' restarts the level. The level also restarts when the ball goes off screen. Player 1 is controlled with W and S, Player 2 is controlled with Up and Down arrows.

Most of these things are hard coded and should be abstracted later.