Implement basic collision detection
This collision detection works by changing the sprite green when a collision is active and changing it white when it's not colliding with anything. There's no collision response so the objects still just pass through each other and it's quite sloppy in implementation. Next will be a rewrite using what should be a cleaner implementation that will also make the collision response portion easier.