Fix AI jittering. Closes #17
Fixed the AI jittering by loosening the range of points considered 'correct' to stop on. The jittering was caused by the paddle never being able to stop exactly on the single point the AI was aiming for since the paddle moves at a flat velocity for a certain time frame.